Create README.md
This project explores the integration of Arduino Cloud and Blynk Cloud on a single ESP32 microcontroller. Both platforms have limitations—Arduino Cloud allows 
control of up to 5 variables, and Blynk Cloud provides 5 virtual datastreams—making it rather limiting.

In this tutorial, I share a bare minimum setup that learners can use as a foundation to expand their IoT projects. This is intended for users with prior
experience in Arduino Cloud and Blynk Cloud, aiming to maximize functionality while using the free-tier options.